third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/directive-parsing-05.html

<!DOCTYPE html>
<html>
<head>
<script>
if (window.testRunner) {
  testRunner.dumpAsText();
  testRunner.dumpChildFrames();
}
</script>
</head>
<body>
  <p>
    Directives starting with an invalid character should be logged and ignored.
  </p>
  <iframe src="http://127.0.0.1:8000/security/contentSecurityPolicy/resources/echo-script-src.pl?should_run=yes&q=http://127.0.0.1:8000/security/contentSecurityPolicy/resources/script.js&csp=:script-src%20'none'"></iframe>
</body>
</html>
