third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/directive-parsing-05-expected.txt

CONSOLE ERROR: The Content-Security-Policy directive name ':script-src' contains one or more invalid characters. Only ASCII alphanumeric characters or dashes '-' are allowed in directive names.
Directives starting with an invalid character should be logged and ignored.

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS
